Add skip option for colored text prompt
When a <span> with color is detected in the build script, a prompt is triggered for
the user to select a Color from the database when the color value is ambiguous. However,
there was no option to skip when the color did not exist at all causing the build
script to be stuck in a loop or require exiting the build process.