Only include active Character names for each Chapter
Each Character can have a "first_chapter_ref" field that references the Chapter the
Character was first mentioned. Character names/aliases are now filtered to only
pass in Characters that are referenced before or during the current Chapter in the
build process.

Note: The "first_chapter_ref" field can be `None` if no URL was provided in the wiki
metadata, or if the reference uses an incorrect or old chapter URL. Any Characters
with `None` in the "first_chapter_ref" field are included for every chapter of the
build process to ensure they are not missed.

This speeds up the build process since it significantly reduces the number of Character
regex expressions that must be compiled, especially for the early chapters before many
of the characters are introduced.