Add build checks for bracketed Character names and Aliases
These checks ignore any brackected characters as they are processed  just prior to the
manual RefType selection prompt to limit how often this check is procced. They are
needed to limit the amount of manual intervention required when building chapters with
many bracketed characters such as "Interlude - The Call" and some later chapters.