feat: add strict mode for TextRef search
Strict mode only applies to the RefType query, not the text search query since
it would require users to write out the entire matching line to return any
results. This features is mainly useful when isolating particular RefTypes. For
instance searching for the character "Ad" would return many other characters
with "Ad" in the name without the strict option.