style: fix table stat_link styles
Firefox and Chrome are inconsistent when rendering the height of cells
in tables and thus cause some differences when rendering the stat_link.html
template in tables. This forces the contents of the table cells to at least
centered vertically. Ideally it would also fill the hover background on hover
with the height: 100% (h-full) but Firefox and Chrome won't both do it with
the same styles, so I've comprimised by getting Chrome to display the way
I want and Firefox will just be centered vertically with a default height
for the button/link.

Fixes issue #33