refactor: parse chapters into partitions for robustness
This refactors a large part of the chapter parsing to partition the
chapter into three parts:
1. pre-note
2. chapter text
3. author's note(s)

These may in some (rare) cases overlap, so partitioning the chapter
by line ranges over the text make it simpler to prevent any duplication
of lines in the parsed output and identify multiple author's notes.

This process is certainly slower since it's not entirely an interative
approach with just one pass like before, but I think it's more robust
this way and should be easier to adjust in the future if needed.

Also adds a few new tests for pre-note and author's note detection.