build: order alias checks by length
When generating TextRefs, the RefType name and its aliases need to be
checked in the order of their length, because some aliases are a superset
of the primary name of the given Character. Otherwise, an Alias that
contains the Character.name would never be matched, which while still
being accurate is not ideal, since the TextRef highlighting would not
contain the full alias.