fix: RefType table data
RefType tables rendered on the sub pages for RefTypes need to use the
strict mode for displaying the correct data. Without it, the results
may include RefTypes with a name that is a subset of other RefTypes
e.g. The character Ad (Captain Ad) includes many other reftypes containing
"ad" in the name.