fix: add required favicon.ico
Insanely, when using the ManifestStaticFilesStorage backend in Django,
if a mapping to that static file doesn't exist in the manifest, then
Django return a server error and crashes the application. So a favicon.ico
is necessary since browsers pretty much always look for it even though
there's already a favicon.svg available.