perf: limit table size to 1000 rows
Showing all rows on extremely large datasets kills render times especially
for tables with chapter links and other templates. This enables the user
to still expand tables to a large number of rows for easier browsing, but
prevents large tables like the Erin Solstice mentions tables from
expanding to much to load in a reasonable amount of time.