admin: add select Model Django admin action
To properly extend the built-in admin site it's necessary to extend
the AdminSite instance, see `StatsAdminSite`.

This adds a new form for admin actions that need to select an item
from a Model's objects. Unfortunately, the `AutocompleteMixin` and
thus the `AutocompleteSelect` widget are tighly coupled with an
implementation that requires a foreign key attribute to be used
to identify the selectable objects.

To work around this, I added a wrapper view (`SelectForeignModelView`)
which takes a `base_model`, `qs_model`, and `field` which can be
used to determine a model and it's foreign key attribute to use
the `AutocompleteSelect` widget, but also serves as a hacky solution
to populate the widget as long as a Model (`base_model`) with a
foreign key to the desired Model (`qs_model`) exists even if the
`base_model` parameters doesn't match the items selected in the admin
dashboard for the specified admin action.

This also refactors the RefType merge actions to use the new view.

It's also important to note that, for these templates to render
properly with the Admin site styles, the associated Media objects
need to be passed into the templates e.g. with `{{ form.media }}`
for the new `SelectForeignModelForm`.