perf: add django-minify-html
Adds minify-html in the middleware to minify the HTML responses.
Specifically, I'm adding this to help reduce the size of tables pages
with _many_ rows, but also to improve page load times and bandwidth
usage overall.

Note that the version of minify-html needs to be pinned to v0.15.0 since
v0.16.0 breaks minification for simple script.
For details see: https://github.com/wilsonzlin/minify-html/issues/242