fix: render error message for RefType gallery charts without enough data
The convention is to return None if a chart geneartion function is unable
to render a meaningful chart with the provided data. Specifically this
currently applies to the "mentions by book" chart.

Now the message "Not enough data" will be rendered in place of the the chart
thumbnail if a RefType chart thumbnail image was not found which _should_
indicate a thumbnail was not generated due to insufficient data.