fix: search page refs by chapter and pagination
- Fix error for htmx pagination, since form.is_valid wasn't getting
called which is needed to populate form.cleaned_data.
- Also removes strict mode from search page, there may be a need for strict
search configuration later, but I think quoted text searches should be enough.
They just need be documented on the search demo page. Additionally
once a filterable RefType dropdown is added, there won't be a need for
the strict mode in the RefType filtering anyway.