fix: text search phrase highlighting
The default Postgres headline highlighting for phrases is
pretty bad. It highlights individual words that are included
in a "phrase" and also doesn't fully encapsulate the actual
phrase highlight meaning there are gaps from the spaces between
words which also looks bad.

This fix applies manual highlighting on the application side so that
we can have more control of the results.