perf: generate interactive charts as static HTML
This improves load times quite a bit for interactive charts. Especially
for the expensive ones like the "Unique Characters per Chapter" chart.

With the previously added threshold on chart generation, this only costs
about 300MB on disk, but doesn't add any meaningful amount of time to
the chart generation script. I think it's a reasonable tradeoff. Now
the only major interaction where the users will hit the database are for
text search.

Note interactive charts with a chapter filter must still be generated
dynamically, but this at least improve response times on the default
interactive charts.